# apidocs.json
{
   "swagger": "2.0",
   "info": {
     "title": "Swagger book api",
     "description": "panduan untuk mengakses data buku dengan restful API",
     "version": "1.0"
   },
   "produces": ["application/json"],
   "host": "localhost:3000",
   "basePath": "",
   "paths": {
     "/books": {
       "get": {
         "x-swagger-router-controller": "routes",
         "operationId": "swagTest",
         "tags": ["/books"],
         "description": "untuk mendapatkan semua data buku",
         "parameters": [],
         "responses": {}
       }
     },
     "/books/id/{id}": {
      "get": {
        "x-swagger-router-controller": "middleware-name1",
        "operationId": "swagTest",
        "tags": ["/books/id/"],
        "description": "untuk mendapatkan satu data buku berdasarkan id",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "need book id to get specific book"
          }
        ],
        "responses": {}
      }
    },

     "/books/tambah": {
      "post": {
        "x-swagger-router-controller": "middleware-name1",
        "operationId": "swagTest",
        "tags": "/books/tambah",
        "description": "menambah data buku baru",
        "parameters": [],
        "responses": {}
      }
    },
    "/books/ubah/{id}": {
      "put": {
        "x-swagger-router-controller": "middleware-name1",
        "operationId": "swagTest",
        "tags": ["/books/ubah"],
        "description": "untuk mengubah satu data buku berdasarkan id buku",
        "parameters": [],
        "responses": {}
      }
    },
    "/books/hapus/{:id}": {
      "delete": {
        "x-swagger-router-controller": "middleware-name1",
        "operationId": "swagTest",
        "tags": ["/books/hapus"],
        "description": "untuk menghapus satu data buku berdasarkan id buku",
        "parameters": [
          {
          "namaBuku": "namabuku1",
          "penerbit": "penerbit1",
          "pengarang": "pengarang1"
          }
        ],
        "responses": {}
      }
    }
   }
}